+++ Assets/Scripts/enemyMove.cs
@@ -12,7 +12,7 @@ public class enemyMove : MonoBehaviour
     public HealthBar healthbar;
     //public Vector3 healthbarOffset;
     public float attackSpeed = 1f;
-    public float attackDamage = 1f;
+    public int attackDamage = 1;
     public bool canAttack = true;
     public AudioSource attackSound;
     public float attackDistance = 1f;
